Create ASCIII code
Write a program that prompts the user to enter a phrase and then prints out the ASCII code of each character in the phrase.
A sample run of your program should look like:

   Enter a phrase:  I love Python!
   In ASCII: 
   73
   32
   108
   111
   118
   101
   32
   80
   121
   116
   104
   111
   110
   33
Hint: If c is a character, ord(c) returns its ASCII code. For example, if c is ‘I’, then ord(c) returns 73. See Lab 2.