Create Code words
(The cipher disk above shifts ‘A’ to ‘N’, ‘B’ to ‘O’, … ‘Z’ to ‘M’, or a shift of 13. From secretcodebreaker.com.) Write a program that prompts the user to enter a word and then prints out the word with each letter shifted right by 13. That is, ‘a’ becomes ‘n’, ‘b’ becomes ‘o’, … ‘y’ becomes ‘l’, and ‘z’ becomes ‘m’.

Assume that all inputted words are in lower case letters: ‘a’,…,’z’.

A sample run of your program should look like:

   Enter a word: zebra 
   Your word in code is: 
   mroen
Hint: See the example programs from Lecture 2.