Create Binary to Decimal
Write a program that implements the pseudocode below. Your program should ask the user for a binary number and print out the corresponding decimal number.

Ask user for input, and store in the string, binString.

   Set decNum = 0.
   For each c in binString,
   Set n to be int(c)
   Double decNum and add n to it (decNum = 2 * decNum + n)
   Print decNum
A sample run of the program:

   Enter binary number: 10
   Your number in decimal is 2
And another sample run of the program:

   Enter binary number: 1111
   Your number in decimal is 15
Note: here’s a quick review of binary numbers.