Create Turtle String
The program turtleString.py takes a string as input and uses that string to control what the turtle draws on the screen (inspired by code.org’s graph paper programming). Currently, the program processes the following commands:

‘F’: moves the turtle forward 50 steps
‘L’: turns the turtle 90 degrees to the left
‘R’: turns the turtle 90 degrees to the right
’^’: lifts the pen
‘v’: lowers the pen
For example, if the user enters the string “FLFLFLFL^FFFvFLFLFLFL”, the turtle would move forward and then turn left. It repeats this 4 times, drawing a square. Next, it lifts the pen and move forward 3, puts the pen back down and draws another square.

Modify this program to allow the user also to specify with the following symbols:

‘B’: moves the turtle backward 50 steps
‘S’: makes the turtle stamp
‘l’: turns the turtle 45 degrees to the left
‘r’: turns the turtle 45 degrees to the right
‘p’: change the pen color to purple