Create Plurals
Write a program that asks the user for a list of nouns (separated by spaces) and approximates the fraction that are plural by counting the fraction that end in s. Your program should output the total number of words and the fraction that end in s. You may assume that words are separated by spaces (and ignore the possibility of tabs and punctuation between words.)